Create LICENSE
Copyright © 2025 Devin B. Royal. All Rights Reserved.

All rights reserved. No part of this software, including but not limited to the algorithms, source code, documentation, or associated materials, may be reproduced, distributed, modified, or used in any form or by any means, including photocopying, recording, or other electronic or mechanical methods, without the prior written permission of Devin B. Royal, except as permitted by applicable law.

This software is provided "as is" without warranty of any kind, express or implied, including but not limited to the warranties of merchantability, fitness for a particular purpose, or non-infringement. In no event shall Devin B. Royal be liable for any claim, damages, or other liability arising from the use of this software.

For permissions or inquiries, contact Devin B. Royal (contact details TBD).